Fix issue with shared_ptr, Python, pybind11, etc.
